Update README.md
In this analysis, I filtered a dataset of Kickstarter projects in order to find the rate of success by financial goal within the category of "theater" and subcategory of "plays". The first conclusion that can be reached from this dataset is the majority of Kickstarter projects within the "plays" subcategory set their target goal below $20,000. Because of this. one of the limitations of this dataset is that the percentage of successful/failed/canceled categories in the first Line Graph does not account for the number of projects, and therefore this data may not be accurate. 

Another conclusion that can be drawn is that Kickstarters focused on raising money for plays were never canceled prior to finishing. This deserves further exploration; we should now compare the cancellation rate with other subcategories using a stacked Bar Graph. 

The third conclusion that can be reached is that the majority of Kickstarters rasing money for plays were successful. This also merits further analysis; we should compare the success rate using our Pivot Table to filter across categories and subcategories. Then, we can use that to compare with our stacked bar graph.